+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Support --locale command line option
+* Support --preferences command line option
 
 # 0.1.0